--- src/axios-orders.js
@@ -1,7 +1,0 @@
-import axios from 'axios';
-
-const instance = axios.create({
-  baseURL: 'https://react-guide-burger.firebaseio.com/'
-});
-
-export default instance;
