+++ src/containers/checkout/contact-data/contact-data.js
@@ -2,7 +2,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 
-import axios from '../../../axios-orders';
+import axios from '../../../axios-service';
 import withErrorHandler from '../../../hoc/errorHandler';
 import * as actions from '../../../store/actions/index';
 import { createArrayOfFormElements, updateValidatedForm, isFormValid } from '../../../utils/index';
