+++ src/containers/orders/orders.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-unused-vars */
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import axios from '../../axios-orders';
+import axios from '../../axios-service';
 import withErrorHandler from '../../hoc/errorHandler';
 import * as actions from '../../store/actions/index';
